# refactor: use in-memory cache to refactor disk cache