--- a/foyer-storage-bench/src/main.rs
+++ b/foyer-storage-bench/src/main.rs
@@ -34,16 +34,17 @@ use analyze::{analyze, monitor, Metrics};
 use clap::Parser;
 use export::MetricsExporter;
 use foyer_common::code::{StorageKey, StorageValue};
-use foyer_intrusive::eviction::lfu::LfuConfig;
+use foyer_memory::LfuConfig;
 use foyer_storage::{
     admission::{rated_ticket::RatedTicketAdmissionPolicy, AdmissionPolicy},
     compress::Compression,
     device::fs::FsDeviceConfig,
     error::Result,
+    region_manager::EvictionConfg,
     reinsertion::{rated_ticket::RatedTicketReinsertionPolicy, ReinsertionPolicy},
     runtime::{RuntimeConfig, RuntimeStore, RuntimeStoreConfig, RuntimeStoreWriter},
     storage::{AsyncStorageExt, Storage, StorageExt, StorageWriter},
-    store::{LfuFsStoreConfig, Store, StoreConfig, StoreWriter},
+    store::{FsStoreConfig, Store, StoreConfig, StoreWriter},
 };
 use futures::future::join_all;
 use itertools::Itertools;
@@ -537,10 +538,12 @@ async fn main() {
     let metrics_dump_start = metrics.dump();
     let time = Instant::now();
 
-    let eviction_config = LfuConfig {
-        window_to_cache_size_ratio: 1,
-        tiny_lru_capacity_ratio: 0.01,
-    };
+    let eviction_config = EvictionConfg::Lfu(LfuConfig {
+        window_capacity_ratio: 0.1,
+        protected_capacity_ratio: 0.8,
+        cmsketch_eps: 0.001,
+        cmsketch_confidence: 0.9,
+    });
 
     let device_config = FsDeviceConfig {
         dir: PathBuf::from(&args.dir),
@@ -573,7 +576,7 @@ async fn main() {
         .try_into()
         .expect("unsupported compression algorithm");
 
-    let config = LfuFsStoreConfig {
+    let config = FsStoreConfig {
         name: "".to_string(),
         eviction_config,
         device_config,