@MrCroxx 2024-04-15
## benchmark

### command

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-storage-bench --dir /p44pro/foyer --capacity 204800 --region-size 64 --lookup-range 10000 --flushers 12 --reclaimers 12 --time 180 --writers 64 --w-rate 8 --ticket-insert-rate-limit 500 --readers 1024 --r-rate 1 --runtime --metrics --compression none --io-size 262144 --distribution zipf --distribution-zipf-s 0.8
```

### main

```
Total:
disk total iops: 3851.0
disk total throughput: 486.0 MiB/s
disk read iops: 379.4
disk read throughput: 25.2 MiB/s
disk write iops: 3471.6
disk write throughput: 460.8 MiB/s
insert iops: 6937.1/s
insert throughput: 433.6 MiB/s
insert lat p50: 35us
insert lat p90: 81us
insert lat p99: 447us
insert lat p999: 991us
insert lat p9999: 1943us
insert lat p99999: 3087us
insert lat pmax: 4063us
get iops: 16363.2/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 1039us
get hit lat p90: 2143us
get hit lat p99: 3167us
get hit lat p999: 4831us
get hit lat p9999: 8895us
get hit lat p99999: 14271us
get hit lat pmax: 15551us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```

### PR

```
Total:
disk total iops: 3855.3
disk total throughput: 485.3 MiB/s
disk read iops: 397.9
disk read throughput: 26.4 MiB/s
disk write iops: 3457.4
disk write throughput: 458.8 MiB/s
insert iops: 6908.0/s
insert throughput: 431.8 MiB/s
insert lat p50: 35us
insert lat p90: 86us
insert lat p99: 711us
insert lat p999: 1407us
insert lat p9999: 2447us
insert lat p99999: 3375us
insert lat pmax: 4575us
get iops: 16361.1/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 1607us
get hit lat p90: 3183us
get hit lat p99: 4223us
get hit lat p999: 5887us
get hit lat p9999: 10175us
get hit lat p99999: 16511us
get hit lat pmax: 17279us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```